Fix database, add MongoDB cloud
This is the cloud version of the database. You must use my password and input string to access. I will see if I can add you as users to this schema.

Also, this is in the mgmt_db, not the wrightandmorrow one. I noticed after I transferred everything. Will create pull request to fix at later data